refactor(pathpida): remove aspida from module
