+++ packages/aspida-mock/src/cli/index.ts
@@ -2,7 +2,7 @@ import minimist from 'minimist'
 import rimraf from 'rimraf'
 import getBaseConfig, { BaseConfig } from 'aspida/dist/getConfig'
 import write from './writeRouteFile'
-import watch from 'aspida/dist/watchInputDir'
+import watch from 'aspida/dist/cli/watchInputDir'
 import { options } from 'aspida/dist/cli'
 import { Build, Watch, CommandToBuild } from './build'
 import { Command, nullCommand } from 'aspida/dist/cli/command'
