+++ packages/aspida/src/cli/index.ts
@@ -2,7 +2,7 @@ import fs from 'fs'
 import minimist from 'minimist'
 import getConfig, { BaseConfig } from '../getConfig'
 import write from '../writeRouteFile'
-import watch from '../watchInputDir'
+import watch from './watchInputDir'
 import { Build, Watch, CommandToBuild } from './build'
 import { Command, nullCommand } from './command'
 import { version as versionCommand } from './version'
