+++ packages/openapi2aspida/src/cli.ts
@@ -3,7 +3,7 @@ import minimist from 'minimist'
 import rimraf from 'rimraf'
 import getConfig, { Config } from './getConfig'
 import write from './writeRouteFile'
-import watch from 'aspida/dist/watchInputDir'
+import watch from 'aspida/dist/cli/watchInputDir'
 import { options } from 'aspida/dist/cli'
 import { Build, Watch, CommandToBuild } from './cli/build'
 import { Command, nullCommand } from './cli/command'
