+++ packages/pathpida/package.json
@@ -1,6 +1,6 @@
 {
   "name": "pathpida",
-  "version": "0.3.3",
+  "version": "0.3.4",
   "description": "TypeScript friendly path generator for Next.js or Nuxt.js",
   "author": "m-mitsuhide <m.mitsuhide@amatelus.com>",
   "license": "MIT",
@@ -21,6 +21,7 @@
     "path"
   ],
   "dependencies": {
-    "aspida": "^0.14.3"
+    "chokidar": "^3.3.1",
+    "minimist": "^1.2.3"
   }
 }
