+++ packages/pathpida/src/cli/command.ts
@@ -1,0 +1,9 @@
+export interface Command {
+  exec(): void
+}
+
+export const nullCommand: Command = {
+  exec() {
+    // nothing to do
+  }
+}
