+++ packages/pathpida/src/cli/version.ts
@@ -1,0 +1,8 @@
+import path from 'path'
+import { Command } from './command'
+
+export const version: Command = {
+  exec() {
+    console.log(`v${require(path.join(process.cwd(), 'package.json')).version}`)
+  }
+}
